libbt-vendor: Fix Samsung patchfile detection.
This is totally broken. If the chipset is bcm4530 then it added e.g.
_wisol to it and looked for bcm4530_wisol.hcd. The firmware files are
named e.g. bcm4350_V0301.0636_wisol.hcd. What we need to do if a special
flavor is needed to search for the string in the name of the file we are
currently inspecting.

Change-Id: If804de56064cf8993a32af98c7ec39126ebd349a
(cherry picked from commit c6884d2d4a6e52a2fb2c44cb787dd5f2fe3a2991)